bugfix: abort if imfile reads file line of more than 64KiB
Thanks to Peter Eisentraut for reporting and analysing this problem.
bug tracker: http://bugzilla.adiscon.com/show_bug.cgi?id=221

+++ runtime/stringbuf.c
@@ -156,7 +156,7 @@ rsRetVal
 rsCStrExtendBuf(cstr_t *pThis, size_t iMinNeeded)
 {
 	uchar *pNewBuf;
-	unsigned short iNewSize;
+	size_t iNewSize;
 	DEFiRet;
 
 	/* first compute the new size needed */
